Start up redis server before each suite
Conflicts:
	src/test/java/redis/clients/johm/JOhmTestBase.java
